+++ test/fake_logger.ts
@@ -5,4 +5,5 @@ export class FakeLogger implements ILogger {
   debug = sinon.fake()
   error = sinon.fake()
   warn = sinon.fake()
+  info = sinon.fake()
 }